Change entity name from TrackingCode to TrackingEvent
Tracking code is not the best name for the entity because it does not represent only a tracking code but all the events for a certain tracking code.

The TrackingEvent now have a code attribute to store the code that an event refers to

It is also a future possibility to create the TrackingCode entity again and add to it an events attribute. For now it was decided to move on with this solution.